+++ nextstep-mvc/src/test/java/samples/MyController.java
@@ -1,6 +1,7 @@
-package nextstep.mvc.tobe;
+package samples;
 
 import nextstep.db.DataBase;
+import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
