+++ nextstep-mvc/src/test/java/samples/TestUserController.java
@@ -1,5 +1,6 @@
-package nextstep.mvc.tobe;
+package samples;
 
+import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.web.annotation.PathVariable;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
